Removed old nveto_pulses and added nveto_hitlets
